requirements.txt
torch
transformers
datasets
bitsandbytes
pandas
numpy
pillow
bert-score
